refactor: auth view model 적용 & 불필요 로직 제거
