+++ app/src/main/java/org/sopt/sample/presentation/auth/view/SignInActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.presentation.signin.view
+package org.sopt.sample.presentation.auth.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,16 +6,15 @@ import androidx.activity.viewModels
 import org.sopt.sample.R
 import org.sopt.sample.databinding.ActivitySignInBinding
 import org.sopt.sample.presentation.MainActivity
+import org.sopt.sample.presentation.auth.viewmodel.AuthViewModel
 import org.sopt.sample.presentation.common.ViewModelFactory
-import org.sopt.sample.presentation.signin.viewmodel.SignInViewModel
-import org.sopt.sample.presentation.signup.view.SignUpActivity
 import org.sopt.sample.util.EventObserver
 import org.sopt.sample.util.binding.BindingActivity
 import org.sopt.sample.util.extension.hideKeyboard
 import org.sopt.sample.util.extension.showToast
 
 class SignInActivity : BindingActivity<ActivitySignInBinding>(R.layout.activity_sign_in) {
-    private val viewModel: SignInViewModel by viewModels { ViewModelFactory() }
+    private val viewModel: AuthViewModel by viewModels { ViewModelFactory() }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,35 +26,34 @@ class SignInActivity : BindingActivity<ActivitySignInBinding>(R.layout.activity_
     }
 
     private fun setObservers() {
-        viewModel.signInEvent.observe(
-            this, EventObserver { isSuccess ->
-                if (isSuccess) {
-                    showToast(getString(R.string.success_sign_in))
-                    startMainActivity()
-                } else {
-                    showToast(getString(R.string.failure_sign_in))
+        with(viewModel) {
+            authEvent.observe(
+                this@SignInActivity, EventObserver { isSuccess ->
+                    if (isSuccess) {
+                        showToast(getString(R.string.success_sign_in))
+                        startMainActivity()
+                    } else {
+                        showToast(getString(R.string.failure_sign_in))
+                    }
                 }
+            )
+            isValidEmail.observe(
+                this@SignInActivity
+            ) { isValid ->
+                binding.signInIdTil.error =
+                    if (!isValid) getString(R.string.invalid_email) else null
             }
-        )
-        viewModel.isValidEmail.observe(
-            this
-        ) { isValid ->
-            binding.signInIdTil.error =
-                if (!isValid) getString(R.string.invalid_email) else null
-        }
-        viewModel.isValidPassword.observe(
-            this
-        ) { isValid ->
-            binding.signInPasswordTil.error =
-                if (!isValid) getString(R.string.invalid_password) else null
+            isValidPassword.observe(
+                this@SignInActivity
+            ) { isValid ->
+                binding.signInPasswordTil.error =
+                    if (!isValid) getString(R.string.invalid_password) else null
+            }
         }
     }
 
     private fun setOnClickListener() {
         with(binding) {
-            signInSignInBtn.setOnClickListener {
-                this@SignInActivity.viewModel.signIn()
-            }
             signInSignUpBtn.setOnClickListener {
                 startSignUpActivity()
             }
